Explicitly cast thread ID to int64 to avoid compiler error
FBSD's gcc (and possibly others) doesn't like an implicit
conversion here.  Making it explicit.

+++ src/runtime/eval/debugger/cmd/cmd_interrupt.h
@@ -35,7 +35,7 @@ public:
       : DebuggerCommand(KindOfInterrupt),
         m_interrupt(interrupt), m_program(program ? program : ""),
         m_site(site), m_pendingJump(false) {
-    m_threadId = Process::GetThreadId();
+    m_threadId = (int64)Process::GetThreadId();
     if (error) m_errorMsg = error;
   }
